requirements.txt
httpx==0.24.1
numpy==1.24.4
opencv-python==4.8.0.76
Pillow==10.0.0
